pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "procguard"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
    "alembic==1.13.1",
    "sqlalchemy==2.0.25",
    "psycopg2-binary==2.9.9",
    "pg8000",
    "python-dotenv",
    "openai",
    "fastapi",
    "uvicorn"
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2"
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
filterwarnings = [
  "error::DeprecationWarning"
]
